Add files via upload
The goal of the project is to develop a command-line utility that uses Python and the OpenWeatherMap API to receive a city's name and return the current weather prediction. Anyone who needs to rapidly check the weather in a specific city, such as travellers, weather lovers, or anyone who has to organise their day based on the weather, will find the application to be helpful. Error handling will be a part of the solution to deal with API problems and print useful error messages to the user. GitHub Copilot will be used to help with a number of development-related tasks, including using the API, processing data, and handling errors.​